jetson: add preset cmdline args
Nvidia passes in a metric ton of cmdline arguments to their kernels.  As a
matter of fact they go over the 255 char limit supplied in the boot image
header.

To handle the char limit as well as clean up the kernel configs, let's place
all of the predefined cmdline arguments in the bootloader where they will be
appended during the boot.

Signed-off-by: Michael Scott <michael.scott@linaro.org>